# core/vm: set stack capacity to 1024 by default