Also adds benchmarks for push/pop.

benchmarks:

Current default stack capacity:
```
> go test -bench=Stack -benchtime=1024x
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm
BenchmarkStackPush-4   	    1024	        10.7 ns/op
BenchmarkStackPop-4    	    1024	         1.28 ns/op
PASS
ok  	github.com/ethereum/go-ethereum/core/vm	1.861s
```

with default stack capacity of 1024:
```
> go test -bench=Stack -benchtime=1024x
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm
BenchmarkStackPush-4   	    1024	         3.50 ns/op
BenchmarkStackPop-4    	    1024	         1.20 ns/op
PASS
ok  	github.com/ethereum/go-ethereum/core/vm	1.890s
```