@holiman 2020-10-17
See https://github.com/ethereum/go-ethereum/pull/21222 -- to see why we recently changed this from 1024 to 16